+++ src/androidMain/kotlin/uz/yalla/platform/button/NativeCircleIconButton.android.kt
@@ -1,19 +1,21 @@
 package uz.yalla.platform.button
 
 import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.IconButtonDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.unit.dp
+import org.jetbrains.compose.resources.painterResource
 import uz.yalla.design.theme.System
 import uz.yalla.platform.model.IconType
-import uz.yalla.platform.toImageVector
+import uz.yalla.platform.toDrawableResource
 
 @Composable
 actual fun NativeCircleIconButton(
@@ -34,10 +36,10 @@ actual fun NativeCircleIconButton(
             containerColor = System.color.backgroundBase
         )
     ) {
-        Icon(
-            imageVector = iconType.toImageVector(),
+        Image(
+            painter = painterResource(iconType.toDrawableResource()),
             contentDescription = null,
-            tint = System.color.iconBase
+            colorFilter = ColorFilter.tint(System.color.iconBase)
         )
     }
 }